thumbnail있을때 s3에서 삭제해야한다.

+++ src/article/article.delete.controller.js
@@ -7,15 +7,18 @@ const s3 = new aws.S3();
 
 exports.delete = async (req, res) => {
   const articleRemove = await query.findArticleById(req.params._id);
-  const params = {
-    Bucket: process.env.S3_BUCKET_NAME,
-    Key: articleRemove.thumbnail
-  }
-  s3.deleteObject(params, (err, data) => {
-    if(err) {
-      console.log(err)
+  // thumbnail이 있을 때 s3에서 사진 삭제
+  if(req.thumbnail) {
+    const params = {
+      Bucket: process.env.S3_BUCKET_NAME,
+      Key: articleRemove.thumbnail
     }
-  })
+    s3.deleteObject(params, (err, data) => {
+      if(err) {
+        console.log(err)
+      }
+    })
+  }
   
   const article = await query.removeArticle({ _id: req.params._id });
   if (!article) {
